+++ node-express/router/api.js
@@ -12,15 +12,12 @@ const bull = require('@es-labs/node/services/mq/bull').get() // bull message que
 const { gcpGetSignedUrl } = require('@es-labs/node/services/gcp')
 const { memoryUpload, storageUpload } = require(APP_PATH + '/common/upload')
 
-const { UPLOAD_STATIC, UPLOAD_MEMORY } = global.CONFIG
+const { UPLOAD_STATIC, UPLOAD_MEMORY, API_PORT, HTTPS_CERTS } = global.CONFIG
 
 const { authUser } = require('@es-labs/node/auth')
 
 function openMissingFile() {
-  fs.readFile('somefile4.txt', (err, data) => {
-    if (err) throw err
-    // console.log(data)
-  })  
+  fs.readFile('somefile4.txt', (err, data) => { if (err) throw err })
 }
 // openMissingFile()
 
@@ -62,7 +59,7 @@ module.exports = express.Router({caseSensitive: true})
     Promise.reject(new Error('unhandled rejection of promise')) // call on.process unhandledRejection - promise rejection, unhandled
   }))
 
-  .get('/healthcheck', (req, res) => res.json({ message: 'OK', app: APP_NAME, environment: process.env.NODE_ENV, version: APP_VERSION }) ) // health check
+  .get('/healthcheck', (req, res) => res.json({ message: 'OK', app: APP_NAME, environment: process.env.NODE_ENV, version: APP_VERSION, port: API_PORT, https: HTTPS_CERTS ? true : false }) ) // health check
 
   .post('/healthcheck', (req, res) => res.json({ message: 'POST OK' }) ) // POST health check